# backend/dpmain/internal/app/domains/apimodel/response/account_response.go
package response

import "time"

// AccountResponse 账号响应
type AccountResponse struct {
	ID        int64     `json:"id" example:"1"`
	Name      string    `json:"name" example:"John Doe"`
	Email     string    `json:"email" example:"john@example.com"`
	CreatedAt time.Time `json:"created_at" example:"2024-01-01T00:00:00Z"`
}